refactor interface routine for client and server interaction: method_connection_cl and method_connection_sr
