+++ cs/CMakeLists.txt
@@ -284,6 +284,7 @@ set(JSP_HEADERS
   )
 
 set(METHOD_SOURCES
+  ${METHOD_DIR}/method_connection_cl.cpp
   ${METHOD_DIR}/method_callback.cpp
   ${METHOD_DIR}/method_def.cpp
   ${METHOD_DIR}/method_error.cpp
