+++ sa/CMakeLists.txt
@@ -326,7 +326,8 @@ set(JSP_HEADERS
 
 set(METHOD_SOURCES
   ${METHOD_DIR}/method_callback.cpp
-  ${METHOD_DIR}/method_connection.cpp
+  ${METHOD_DIR}/method_connection_cl.cpp
+  ${METHOD_DIR}/method_connection_sr.cpp
   ${METHOD_DIR}/method_def.cpp
   ${METHOD_DIR}/method_error.cpp
   ${METHOD_DIR}/method_scan.cpp
